ciclope_mitologias_verbales/ciclope_en_siete_capas/requirements.txt
perplexity>=0.1.0
python-dotenv>=0.19.0
requests>=2.31.0
pytest>=7.0.0
pytest-mock>=3.10.0
openai>=1.0.0

